Separated Ddb exe abstraction from DdbStorage
