+++ Registry.Ports/DroneDB/IDdbStorage.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+using Registry.Ports.DroneDB.Models;
+
+namespace Registry.Ports.DroneDB
+{
+    /// <summary>
+    /// Abstracts the drone db operations
+    /// </summary>
+    public interface IDdbStorage : IDisposable
+    {
+        // public DdbObject GetObjectInfo(int id);
+        IEnumerable<DdbObject> Search(string path);
+        void Add(string path, byte[] data);
+        void Remove(string path);
+    }
+}
