+++ Registry.Web/Services/Adapters/DdbFactory.cs
@@ -25,15 +25,15 @@ namespace Registry.Web.Services.Adapters
             _settings = settings.Value;
         }
 
-        // NOTE: This logic is separated from the managers classes because it is used in multiple places and it could be subject to change
+        // NOTE: This logic is separated from the manager classes because it is used in multiple places and it could be subject to change
 
-        public IDdb GetDdb(string orgId, string dsId)
+        public IDdbStorage GetDdb(string orgId, string dsId)
         {
             var ddbPath = Path.Combine(_settings.DdbStoragePath, orgId, dsId, DefaultDdbFolder, DefaultDdbSqliteName);
 
             _logger.LogInformation($"Opening ddb in '{ddbPath}'");
 
-            var ddb = new Ddb(ddbPath, Path.Combine(_settings.DdbPath, DdbPackageProvider.GetExeName()));
+            var ddb = new DdbStorage(ddbPath, new Ddb(Path.Combine(_settings.DdbPath, DdbPackageProvider.GetExeName())));
 
             var res = ddb.Database.EnsureCreated();
